Database editor reworked, added Modals Dialogs and Popup JS Library, slight color adjustments to dark template, PHP View models and CPanel integramtions, PanelView & TransactionView libraries introduced

+++ src/System/Personalization/Personalization.php
@@ -12,16 +12,19 @@ abstract class Personalization
 
 	public function register(int $id): bool|null
 	{
-		$result = $this->app->db->query(
-			"INSERT INTO 
+		if ($this->app->user->logged) {
+			$result = $this->app->db->query(
+				"INSERT INTO 
 				user_settings 
 					(usrset_usr_id, usrset_type, usrset_usr_defind_name, usrset_value) 
 			VALUES 
 					({$this->app->user->info->id},{$this->identifier}, $id, 1) 
 			ON DUPLICATE KEY UPDATE 
 				usrset_value = usrset_value + 1;"
-		);
-		return $result ? true : false;
+			);
+			return $result ? true : false;
+		}
+		return false;
 	}
 	public function list(): \Generator
 	{
@@ -30,8 +33,11 @@ abstract class Personalization
 
 	public static function purgeAllPreferences($app): bool
 	{
-		$result = $app->db->query("DELETE FROM user_settings WHERE usrset_usr_id = {$app->user->info->id}");
-		return $result ? true : false;
+		if ($app->user->logged) {
+			$result = $app->db->query("DELETE FROM user_settings WHERE usrset_usr_id = {$app->user->info->id}");
+			return $result ? true : false;
+		}
+		return false;
 	}
 
 }
